Update dependencies and clean up example project
For some reason, the latest interface in the example does NOT work: in
particular, focus and blur events don't happen. Maybe other stuff is
broken too, but since nothing can get focused this prevents anything
from productively happening in the interface.

So, the example is still using old versions of jjm/qasrl. I will have
to figure out what the problem is...